+++ src/main/java/com/example/auction/common/config/AsyncConfig.java
@@ -1,0 +1,17 @@
+package com.example.auction.common.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableAsync;
+
+import java.util.concurrent.Executor;
+import java.util.concurrent.Executors;
+
+@Configuration
+@EnableAsync
+public class AsyncConfig {
+    @Bean(name = "executorWithVT")
+    public Executor asyncExecutorWithVT() {
+        return Executors.newVirtualThreadPerTaskExecutor();
+    }
+}
